Small Area Estimation Non-Parametric based Nadaraya-Watson Kernel and Bootstrap Mean Squared Error Estimators

This Function Gives Small Area Estimation Non-Parametric based Nadaraya-Watson Kernel and Calculates the bootstrap mean squared error estimates

mse_saekernel(X, Y, vardir, bandwidth, B = 1000)

Arguments

X
Auxiliary variable of X

Y
Direct Estimation of Y

vardir
Sampling variances of direct estimators

bandwidth
The kernel bandwidth smoothing parameter

B
Number of bootstrap. Default is 1000

Value

This function returns a list with following objects:

est

a value of Small Area Estimation Non-Parametric based Nadaraya-Watson Kernel estimators

refvar

estimated random effect variance

mse

estimated Bootstrap Mean Squared Error Estimators of Small Area Estimation Non-Parametric based Nadaraya-Watson Kernel

Examples

##load dataset
data(Data_saekernel)

mse_saekernel(X = Data_saekernel$x, Y = Data_saekernel$y,
vardir = Data_saekernel$Vardir, bandwidth = 0.04, B = 1000)

#> $est
#>            [,1]
#>   [1,] 5.046430
#>   [2,] 5.029341
#>   [3,] 4.992263
#>   [4,] 4.980043
#>   [5,] 5.050442
#>   [6,] 4.887401
#>   [7,] 5.043081
#>   [8,] 4.987220
#>   [9,] 4.980274
#>  [10,] 4.934672
#>  [11,] 5.001671
#>  [12,] 5.077403
#>  [13,] 5.024121
#>  [14,] 5.032049
#>  [15,] 4.933460
#>  [16,] 5.055641
#>  [17,] 4.917896
#>  [18,] 4.989176
#>  [19,] 5.044402
#>  [20,] 5.029775
#>  [21,] 5.042856
#>  [22,] 5.071936
#>  [23,] 4.962980
#>  [24,] 4.966874
#>  [25,] 4.906478
#>  [26,] 5.094151
#>  [27,] 5.009970
#>  [28,] 5.054156
#>  [29,] 5.056380
#>  [30,] 5.035698
#>  [31,] 5.076729
#>  [32,] 5.146463
#>  [33,] 5.160613
#>  [34,] 5.121842
#>  [35,] 5.072688
#>  [36,] 5.084106
#>  [37,] 5.048559
#>  [38,] 4.972562
#>  [39,] 5.211547
#>  [40,] 5.065822
#>  [41,] 5.118315
#>  [42,] 5.001015
#>  [43,] 5.029375
#>  [44,] 5.177544
#>  [45,] 5.266243
#>  [46,] 5.317070
#>  [47,] 5.205209
#>  [48,] 5.296312
#>  [49,] 5.446503
#>  [50,] 5.402719
#>  [51,] 5.542108
#>  [52,] 5.665399
#>  [53,] 5.558066
#>  [54,] 5.611203
#>  [55,] 5.540088
#>  [56,] 5.767747
#>  [57,] 5.767574
#>  [58,] 5.927471
#>  [59,] 6.109826
#>  [60,] 5.764549
#>  [61,] 5.931464
#>  [62,] 5.871466
#>  [63,] 6.016694
#>  [64,] 6.047677
#>  [65,] 5.891748
#>  [66,] 5.570403
#>  [67,] 5.607115
#>  [68,] 5.502365
#>  [69,] 5.362446
#>  [70,] 5.094083
#>  [71,] 5.181889
#>  [72,] 5.135589
#>  [73,] 5.011321
#>  [74,] 5.219926
#>  [75,] 5.125200
#>  [76,] 5.102771
#>  [77,] 5.093848
#>  [78,] 4.982674
#>  [79,] 5.058699
#>  [80,] 5.054694
#>  [81,] 5.026431
#>  [82,] 5.091747
#>  [83,] 4.960544
#>  [84,] 5.072411
#>  [85,] 4.989128
#>  [86,] 5.008655
#>  [87,] 4.817225
#>  [88,] 4.644501
#>  [89,] 4.617055
#>  [90,] 4.580638
#>  [91,] 4.411485
#>  [92,] 4.354828
#>  [93,] 4.322076
#>  [94,] 4.092688
#>  [95,] 4.254024
#>  [96,] 4.409327
#>  [97,] 4.665591
#>  [98,] 4.894477
#>  [99,] 4.832266
#> [100,] 4.928012
#> 
#> $refvar
#> [1] 0.08098724
#> 
#> $mse
#>                [,1]
#>   [1,] 5.010774e-04
#>   [2,] 1.378510e-02
#>   [3,] 2.174058e-03
#>   [4,] 1.724104e-02
#>   [5,] 2.289006e-04
#>   [6,] 3.351346e-03
#>   [7,] 1.140762e-03
#>   [8,] 2.575959e-02
#>   [9,] 2.424374e-03
#>  [10,] 1.183144e-03
#>  [11,] 1.766228e-02
#>  [12,] 1.909575e-02
#>  [13,] 1.714409e-02
#>  [14,] 1.719157e-02
#>  [15,] 6.742260e-05
#>  [16,] 2.764664e-03
#>  [17,] 3.205053e-05
#>  [18,] 5.025480e-03
#>  [19,] 5.397515e-02
#>  [20,] 5.953396e-05
#>  [21,] 2.731051e-02
#>  [22,] 2.211802e-02
#>  [23,] 1.381951e-03
#>  [24,] 2.934153e-02
#>  [25,] 2.245367e-03
#>  [26,] 1.633466e-02
#>  [27,] 4.216619e-02
#>  [28,] 1.637646e-02
#>  [29,] 4.771482e-02
#>  [30,] 1.619418e-02
#>  [31,] 3.075081e-02
#>  [32,] 8.979027e-03
#>  [33,] 3.390604e-03
#>  [34,] 1.707976e-02
#>  [35,] 6.707680e-02
#>  [36,] 1.275470e-02
#>  [37,] 1.133788e-02
#>  [38,] 5.820519e-03
#>  [39,] 2.039031e-04
#>  [40,] 2.134208e-03
#>  [41,] 1.546356e-02
#>  [42,] 1.857605e-02
#>  [43,] 4.151621e-03
#>  [44,] 2.662664e-02
#>  [45,] 2.672365e-06
#>  [46,] 3.597185e-03
#>  [47,] 2.934889e-02
#>  [48,] 8.825788e-03
#>  [49,] 9.652483e-03
#>  [50,] 2.175753e-04
#>  [51,] 2.918807e-02
#>  [52,] 3.493109e-03
#>  [53,] 4.715547e-04
#>  [54,] 8.162063e-03
#>  [55,] 1.353507e-04
#>  [56,] 7.704161e-04
#>  [57,] 3.402681e-02
#>  [58,] 1.394449e-03
#>  [59,] 5.273745e-05
#>  [60,] 5.395568e-03
#>  [61,] 2.902969e-03
#>  [62,] 1.106285e-03
#>  [63,] 6.983630e-04
#>  [64,] 2.756925e-03
#>  [65,] 1.230904e-02
#>  [66,] 1.046295e-03
#>  [67,] 4.885320e-03
#>  [68,] 1.828529e-02
#>  [69,] 3.486823e-02
#>  [70,] 7.372812e-06
#>  [71,] 5.142603e-04
#>  [72,] 2.197374e-02
#>  [73,] 1.619984e-05
#>  [74,] 8.934643e-03
#>  [75,] 3.399125e-02
#>  [76,] 3.015925e-03
#>  [77,] 6.661802e-02
#>  [78,] 2.076488e-03
#>  [79,] 5.507841e-04
#>  [80,] 3.901577e-02
#>  [81,] 1.033283e-02
#>  [82,] 2.800651e-05
#>  [83,] 2.204617e-03
#>  [84,] 1.576620e-02
#>  [85,] 8.139363e-03
#>  [86,] 7.187747e-05
#>  [87,] 2.585000e-02
#>  [88,] 1.781190e-03
#>  [89,] 3.651327e-04
#>  [90,] 3.158785e-03
#>  [91,] 3.530549e-03
#>  [92,] 3.488994e-03
#>  [93,] 7.722267e-03
#>  [94,] 1.355696e-03
#>  [95,] 2.009372e-02
#>  [96,] 7.330607e-03
#>  [97,] 7.328912e-03
#>  [98,] 2.097581e-03
#>  [99,] 2.283811e-03
#> [100,] 2.108618e-06
#>